dlt-workshop/tutorial/pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dlt[workspace]>=1.22.1",
    "marimo",
    "ibis-framework[duckdb]",
    "matplotlib",
]
